# imlib: Optimize math ops.