… src/omv/imlib/mathop.c