Optimizes add()/sub()/min()/max()/difference() using cortex-m SIMD so that they work 32-bit at a time. Grayscale operations are faster and RGB565 are faster.

negate() is deleted by PR: https://github.com/openmv/openmv/pull/2139
replace() and blend() will be removed in the future by switching to a draw_image backend. So, they were not optimized.
mul() and div() are removed by PRs: https://github.com/openmv/openmv/pull/2143 and https://github.com/openmv/openmv/pull/2140 respectively.

To save firmware space, I call the binary line op from some of these line ops when the result is the same.

Broke up from: https://github.com/openmv/openmv/pull/2061

...

Performance Numbers (on OpenMV Cam Plus):

VGA Resolution:

add:
* Binary: N/A See b_or performance
* Grayscale: 6968.43us (previous 8225.84) - 18% faster
* RGB565: 16332.4us (previous 18264.0) - 11% faster

sub:
* Binary: 460.541us (previous 10931.5) - 2273% faster
* Grayscale: 6923.17us (previous 8224.25) - 18% faster
* RGB565: 13922.5us (previous 17192.8) - 23% faster

rsub:
* Binary: 443.577us (previous 10527.5) - 2273% faster
* Grayscale: 6899.38us (previous 8191.05) - 18% faster
* RGB565: 13888.0us (previous 15682.7) - 13% faster

min:
* Binary: N/A See b_and performance
* Grayscale: 6231.77us (previous 7560.89) - 21% faster
* RGB565: 15830.7us (previous 18376.8) - 16% faster

max:
* Binary: N/A See b_or performance
* Grayscale: 6232.52us (previous 7562.21) - 21% faster
* RGB565: 15874.9us (previous 18434.9) - 16% faster

difference:
* Binary: N/A See b_xor performance
* Grayscale: 6954.57us (previous 8238.88) - 18% faster
* RGB565: 16258.3us (previous 18605.3) - 14% faster

Grayscale and RGB565 image processing code should increase in performance on faster access to image data. I think the numbers were maxing out on the memory bus bandwidth as the code for processing grayscale images should be 4x faster and the code for processing RGB565 images should be 2x faster.